flux-server/requirements.txt
torch>=2.2.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
peft>=0.11.0
safetensors>=0.4.0
sentencepiece>=0.2.0
protobuf>=4.25.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
Pillow>=10.3.0
huggingface-hub>=0.23.0
gradio>=4.36.0
requests>=2.32.0
pandas>=2.2.0
tqdm>=4.66.0
